+++ src/hdusd/bl_nodes/nodes/shader.py
@@ -15,27 +15,7 @@
 import math
 
 from ..node_parser import NodeParser
-from ...utils import logging
-log = logging.Log(tag='export.bl_nodes.nodes')
-
-
-# INPUTS
-
-class ShaderNodeValue(NodeParser):
-    """ simply return val """
-
-    def export(self):
-        return self.get_output_default()
-
-
-class ShaderNodeRGB(NodeParser):
-    """ simply return val """
-
-    def export(self):
-        return self.get_output_default()
-
-
-# TEXTURES
+from . import log
 
 
 # SHADERS
@@ -226,147 +206,6 @@ class ShaderNodeEmission(NodeParser):
         return result
 
 
-# COLOR
-
-class ShaderNodeInvert(NodeParser):
-    def export(self):
-        fac = self.get_input_value('Fac')
-        color = self.get_input_value('Color')
-
-        return fac.blend(color, 1.0 - color)
-
-
-class ShaderNodeMixRGB(NodeParser):
-
-    def export(self):
-        fac = self.get_input_value('Fac')
-        color1 = self.get_input_value('Color1')
-        color2 = self.get_input_value('Color2')
-
-        # these mix types are copied from cycles OSL
-        blend_type = self.node.blend_type
-
-        if blend_type in ('MIX', 'COLOR'):
-            res = fac.blend(color1, color2)
-
-        elif blend_type == 'ADD':
-            res = fac.blend(color1, color1 + color2)
-
-        elif blend_type == 'MULTIPLY':
-            res = fac.blend(color1, color1 * color2)
-
-        elif blend_type == 'SUBTRACT':
-            res = fac.blend(color1, color1 - color2)
-
-        elif blend_type == 'DIVIDE':
-            res = fac.blend(color1, color1 / color2)
-
-        elif blend_type == 'DIFFERENCE':
-            res = fac.blend(color1, abs(color1 - color2))
-
-        elif blend_type == 'DARKEN':
-            res = fac.blend(color1, color1.min(color2))
-
-        elif blend_type == 'LIGHTEN':
-            res = fac.blend(color1, color1.max(color2))
-
-        elif blend_type == 'VALUE':
-            res = color1
-
-        elif blend_type == 'SCREEN':
-            tm = 1.0 - fac
-            res = 1.0 - (tm + fac * (1.0 - color2)) * (1.0 - color1)
-
-        elif blend_type == 'SOFT_LIGHT':
-            tm = 1.0 - fac
-            scr = 1.0 - (1.0 - color2) * (1.0 - color1)
-            res = tm * color1 + fac * ((1.0 - color1) * color2 * color1 + color1 * scr)
-
-        elif blend_type == 'LINEAR_LIGHT':
-            test_val = color2 > 0.5
-            res = test_val.if_else(color1 + fac * (2.0 * (color2 - 0.5)),
-                                   color1 + fac * (2.0 * color2 - 1.0))
-
-        else:
-            # TODO: support operations SATURATION, HUE, SCREEN, BURN, OVERLAY
-            log.warn("Ignoring unsupported Blend Type", blend_type, self.node, self.material,
-                     "mix will be used")
-            res = fac.blend(color1, color2)
-
-        if self.node.use_clamp:
-            res = res.clamp()
-
-        return res
-
-
-# CONVERTER
-
-class ShaderNodeMath(NodeParser):
-    """ Map Blender operations to MaterialX definitions, see the stdlib_defs.mtlx in MaterialX """
-
-    def export(self):
-        op = self.node.operation
-        in1 = self.get_input_value(0)
-        # single operand operations
-        if op == 'SINE':
-            res = in1.sin()
-        elif op == 'COSINE':
-            res = in1.cos()
-        elif op == 'TANGENT':
-            res = in1.tan()
-        elif op == 'ARCSINE':
-            res = in1.asin()
-        elif op == 'ARCCOSINE':
-            res = in1.acos()
-        elif op == 'ARCTANGENT':
-            res = in1.atan()
-        elif op == 'LOGARITHM':
-            res = in1.log()
-        elif op == 'ABSOLUTE':
-            res = abs(in1)
-        elif op == 'FLOOR':
-            res = in1.floor()
-        elif op == 'FRACT':
-            res = in1 % 1.0
-        elif op == 'CEIL':
-            res = in1.ceil()
-        elif op == 'ROUND':
-            f = in1.floor()
-            res = (in1 % 1.0 < 0.5).if_else(f, f + 1.0)
-
-        else:  # 2-operand operations
-            in2 = self.get_input_value(1)
-
-            if op == 'ADD':
-                res = in1 + in2
-            elif op == 'SUBTRACT':
-                res = in1 - in2
-            elif op == 'MULTIPLY':
-                res = in1 * in2
-            elif op == 'DIVIDE':
-                res = in1 / in2
-            elif op == 'POWER':
-                res = in1 ** in2
-            elif op == 'MINIMUM':
-                res = in1.min(in2)
-            elif op == 'MAXIMUM':
-                res = in1.max(in2)
-
-            else:
-                in3 = self.get_input_value(2)
-
-                if op == 'MULTIPLY_ADD':
-                    res = in1 * in2 + in3
-                else:
-                    log.warn("Unsupported math operation", op)
-                    return None
-
-        if self.node.use_clamp:
-            res = res.clamp()
-
-        return res
-
-
 class ShaderNodeMixShader(NodeParser):
 
     def export(self):
@@ -383,7 +222,7 @@ class ShaderNodeMixShader(NodeParser):
         if shader2 is None:
             return shader1
 
-        result = self.create_node('mix_bsdf', 'surfaceshader', {
+        result = self.create_node('mix', 'BSDF', {
             'fg': shader1,
             'bg': shader2,
             'mix': factor
@@ -409,7 +248,7 @@ class ShaderNodeAddShader(NodeParser):
         if shader2 is None:
             return shader1
 
-        result = self.create_node('add_bsdf', 'surfaceshader', {
+        result = self.create_node('add', 'BSDF', {
             'in1': shader1,
             'in2': shader2
         })
